Fixed NullPointerException bug for ghastly bees
After breeding ghastly bees (possibly others) in the overworld for a while, server crashed with a NullPointerException every time the chunk with the apiary was loaded. Seems that EntityList.createEntityByName() will return null when spawning a ghast in the overworld. Just added a null check to the mob variable before performing any operations on it. If these bees are only to be bred in the Nether adding a check for that would also fix it, but the null check isn't a bad idea :) Oddly enough, after fixing this the ghast spawned just fine.